backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-docx==1.1.2
python-multipart==0.0.9
mammoth==1.8.0
docx2pdf==0.1.8; platform_system == "Windows"
pywin32==306; platform_system == "Windows"
pypdf==4.2.0
pytest==8.3.2
